Add worker thread for hashing and file io
Lots of breaking changes in here to get this to work so the tests no
longer run.  CLI apps other than replay are untested.

The basic idea is to make a deep copy of the create info struct during
vkCreate* and add it to a work queue where the worker thread can
proceed at its own pace.  All hashing and serialization happens on the
worker thread so the main thread can unlock as soon as possible.