Experiment with a trivial stream format.
A ZIP archive will end in a corrupted state unless the application
is very careful to properly deinit everything since we have to write out
a central archive at the end of archive writing.

For all Steam apps I've tested so far (mostly DXVK apps), this never happens,
and our ZIP archive is broken.

This format should be consistently readable even in the face of
corruption since it can be chopped off at any point and still be
readable.