Linux, fossilize-replay: Lower the IO priority
fossilize_replay seems to be mostly CPU bound but still does a lot of
IO which reduces perceived performance of interactive processes. Let's
give up our share of IO in favor of foreground processes.

This commit hints the kernel that our IO is purely background IO. It
may increase the time needed to traverse all data benefiting the user
doing desktop work or playing a game.

The syscall defines are a lazy and shameless copy from the GameMode
project I contributed to it some time ago.

Signed-off-by: Kai Krakow <kai@kaishome.de>