Linux: Use `/proc/self/exe` to execute external replayer
In Linux, we can just `execv()` ourselves without resolving a symlink.
This has the benefit that we always execute the exact same binary even
when it has been changed externally, e.g. by an update process.

In Windows, the executable cannot be replaced while it is running, so
it is safe there.

This has a cosmetic side-effect of showing children as "exe" in tools
like top.

Signed-off-by: Kai Krakow <kai@kaishome.de>