Rewrite how archives are parsed in a multi-process environment.
In order to stave off memory pressure and disk thrashing, we will parse
the archive once, and encode the resulting metadata in a shared memory
block. This metadata block can then be imported in child processes where
we avoid having to scan through the archive altogether. This saves a lot
of startup churn, and for very large archives + many processes, we save
a significant amount of memory as well, as essentially the same
duplicated information would be placed in memory.