Use Linux PSI to dynamically adjust number of active processes.
A common issue which causes grief is that Fossilize replayer has a
tendency to flood the IO caches with data very quickly which negatively
impacts the entire system responsiveness.

To handle this, we study dirty page cache and IO stall metrics to
dynamically adjust number of active processes at any time.

Since this is now essentially doing user scheduling to a certain degree,
we repurpose the stdin sentinel (now control FD) to also handle various control messages,
which the external replayer caller can use to dynamically adjust number
of processes. This can be used to implement a "foreground vs background"
kind of scheme.

Reorients the control FD to Unix DGRAM sockets for simpler implementation
and avoid magically hijacking stdin for no good reason.