Be more robust against inexplicable deadlocks in waitpid().
There are currently some unexplained deadlocks reported for waitpid(),
but this should not be possible. There might be global signal state
unknown to the external replayer which could cause this, but needs
investigation.

Experimentally wait on an FD instead,
whose last reference closing will serve as a pollable waitpid().
Afterwards, perform non-blocking reaps with a short timeout.