Clear out object -> hash references on failure.
If we fail to serialize an object, we run the risk of reusing object
pointers such that a pipeline using a broken layout or similar will end
up pointing to a bogus pipeline layout, which will lead to crashes in
replay later.

To fix this, clear out the hashmap reference if we don't end up updating
the hash map.

This is more likely to happen when immutable samplers are blocked from
serialization, without it, it is extremely rare.