Add --disable-signal-handler for debugging and coredump-ing.
In case of a hard to reproduce driver / compiler crash, it's desirable to
have a way that does not require attaching a debugger to Fossilize to get
a crash dump / stack trace.

By using this option to disable the signal handler, the crashes will be
directly passed on to a coredump handler (e.g. systemd-coredump), which the
developer can use to diagnose the crash after-the-fact.

Tested with robust and non-robust mode, on SEGV, ABRT and timeout-triggered
ABRT.