Remove use of std::async in ApplicationInfoFilter.
It is causing extremely weird crash issues on GCC 32-bit
with -march=pentium4 and -ftree-slp-vectorize.

The use of std::async has not been all that useful here since we block
on the async task almost immediately after creating the task and there is
little to no gain from using it, so just take the opportunity to
simplify the code to parse the JSON serially.