Fix potential deadlock in fgets in master process.
If we SIGSTOP a while it's writing to the buffer, it's possible we can
observe partially written line, which will deadlock in fgets while it's
blocking, waiting for a newline to arrive. It will never come because
the child is stopped. Use a raw file IO approach which is guaranteed to
never block.